split algo, io
