+++ vs/mob.vcxproj
@@ -107,6 +107,7 @@
     <ClCompile Include="..\src\tools\tools.cpp" />
     <ClCompile Include="..\src\utility.cpp" />
     <ClCompile Include="..\src\utility\fs.cpp" />
+    <ClCompile Include="..\src\utility\io.cpp" />
     <ClCompile Include="..\src\utility\string.cpp" />
     <ClCompile Include="..\src\utility\threading.cpp" />
   </ItemGroup>
@@ -123,8 +124,10 @@
     <ClInclude Include="..\src\tasks\tasks.h" />
     <ClInclude Include="..\src\tools\tools.h" />
     <ClInclude Include="..\src\utility.h" />
+    <ClInclude Include="..\src\utility\algo.h" />
     <ClInclude Include="..\src\utility\assert.h" />
     <ClInclude Include="..\src\utility\fs.h" />
+    <ClInclude Include="..\src\utility\io.h" />
     <ClInclude Include="..\src\utility\string.h" />
     <ClInclude Include="..\src\utility\threading.h" />
   </ItemGroup>
