+++ vs/mob.vcxproj.filters
@@ -162,6 +162,9 @@
     <ClCompile Include="..\src\utility\fs.cpp">
       <Filter>src\utility</Filter>
     </ClCompile>
+    <ClCompile Include="..\src\utility\io.cpp">
+      <Filter>src\utility</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\src\pch.h">
@@ -210,6 +213,12 @@
       <Filter>src\utility</Filter>
     </ClInclude>
     <ClInclude Include="..\src\utility\fs.h">
+      <Filter>src\utility</Filter>
+    </ClInclude>
+    <ClInclude Include="..\src\utility\algo.h">
+      <Filter>src\utility</Filter>
+    </ClInclude>
+    <ClInclude Include="..\src\utility\io.h">
       <Filter>src\utility</Filter>
     </ClInclude>
   </ItemGroup>
